+++ R/blur.R
@@ -1,0 +1,2 @@
+utils::adist
+rlang::is_installed